# src/controllers/contacts.js
import * as fs from "node:fs/promises";
import path from "node:path";

import createHttpError from "http-errors";
import {
  createContact,
  deleteContact,
  getAllContacts,
  getContactById,
  patchContact,
} from "../services/contacts.js";

import { getEnvVar } from "../utils/getEnvVar.js";
import { parsePaginationParams } from "../utils/parsePaginationParams.js";
import { parseSortParams } from "../utils/parseSortParams.js";
import { parseFilterParams } from "../utils/parseFilterParams.js";
import { uploadToCloudinary } from "../utils/uploadToCloudinary.js";

// !GET

export const getContactsController = async (req, res) => {
  const { page, perPage } = parsePaginationParams(req.query);
  const { sortBy, sortOrder } = parseSortParams(req.query);
  const filter = parseFilterParams(req.query);
  const userId = req.user._id;

  const contacts = await getAllContacts({
    page,
    perPage,
    sortBy,
    sortOrder,
    filter,
    userId,
  });

  res.status(200).json({
    status: 200,
    message: "Successfully found contacts!",
    data: {
      data: contacts.data,
      page: contacts.page,
      perPage: contacts.perPage,
      totalItems: contacts.totalItems,
      totalPages: contacts.totalPages,
      hasPreviousPage: contacts.hasPreviousPage,
      hasNextPage: contacts.hasNextPage,
    },
  });
};

export const getContactByIdController = async (req, res) => {
  const { contactId } = req.params;
  const userId = req.user._id;

  const contact = await getContactById(contactId, userId);
  console.log(contact);

  if (!contact) {
    throw createHttpError(404, "Contact not found");
  }

  res.status(200).json({
    status: 200,
    message: `Successfully found contact with id ${contactId}!`,
    data: contact,
  });
};

// !POST

export const createContactController = async (req, res) => {
  let photo = null;

  if (req.file?.path) {
    if (getEnvVar("UPLOAD_TO_CLOUDINARY") === "true") {
      const result = await uploadToCloudinary(req.file.path);

      await fs.unlink(req.file.path);

      photo = result.secure_url;
    } else {
      await fs.rename(
        req.file.path,
        path.resolve("src", "uploads", "photos", req.file.filename)
      );

      photo = `http://localhost:3000/photos/${req.file.filename}`;
    }
  }

  const contact = await createContact({
    ...req.body,
    userId: req.user._id,
    photo,
  });

  res.status(201).json({
    status: 201,
    message: "Successfully created a contact!",
    data: contact,
  });
};

// !PATCH

export const patchContactController = async (req, res) => {
  const { contactId } = req.params;
  const photo = req.file;

  let photoUrl;

  if (photo) {
    if (getEnvVar("UPLOAD_TO_CLOUDINARY") === "true") {
      const result = await uploadToCloudinary(photo.path);
      await fs.unlink(photo.path); // удаляем временный файл
      photoUrl = result.secure_url;
    } else {
      const newFileName = `${Date.now()}_${photo.originalname}`;
      const destination = path.resolve("src", "uploads", "photos", newFileName);
      await fs.rename(photo.path, destination);
      photoUrl = `http://localhost:3000/photos/${newFileName}`;
    }
  }
  const result = await patchContact(
    contactId,
    {
      ...req.body,
      ...(photoUrl && { photo: photoUrl }),
    },
    req.user._id
  );

  if (!result) {
    throw createHttpError(404, "Contact not found");
  }

  res.json({
    status: 200,
    message: "Successfully patched a contact!",
    data: result,
  });
};

// !DELETE

export const deleteContactController = async (req, res) => {
  const { contactId } = req.params;
  const contact = await deleteContact(contactId, req.user._id);

  if (!contact) {
    throw createHttpError(404, "Contact not found");
  }

  res.status(204).end();
};